:recycle:: one huge refactor commit
