+++ Quoridor.Console.App/Program.cs
@@ -1,14 +1,10 @@
-﻿using Quoridor.OutputConsole.Input;
-
-namespace Quoridor.OutputConsole.App
+﻿namespace Quoridor.OutputConsole.App
 {
     internal class Program
     {
         private static void Main()
         {
-            ConsoleInput input = new();
-            input.OnStart();
-            input.ReadMove();
+            GameLoop gameLoop = new();
         }
     }
 }
