+++ Quoridor.Console.Input/Quoridor.OutputConsole.Input.csproj
@@ -6,9 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Quoridor.Model\Quoridor.Model.csproj" />
-    <ProjectReference Include="..\Quoridor.View.Console\Quoridor.View.csproj" />
-  </ItemGroup>
-
 </Project>
