+++ Quoridor.View.Console/ViewOutput.cs
@@ -161,5 +161,119 @@ namespace Quoridor.View
         }
 
         #endregion Methods
+
+
+        private const string GreetingMessage = "Hi! Now You are " +
+            "playing Quoridor.\nThe object of the game is to advance " +
+            "your pawn to the opposite edge of the board.\nOn your " +
+            "turn you may either move your pawn or place a wall. " +
+            "\nYou may hinder your opponent with wall placement, " +
+            "but not completely block them off.\nMeanwhile, " +
+            "they are trying to do the same to you.\nThe first " +
+            "pawn to reach the opposite side wins.";
+        private const string NullOrEmptyMessage = "Your input " +
+            "is empty! Try again";
+        private const string HelpMessage = "Here's some tips " +
+            "tips on how to play the game:\n1. start x - start new " +
+            "game, where is the number of real players " +
+            "(1 for one real and one bot. 2 for two real players)" +
+            "\n2. move xy - move Your player to cell xy, for " +
+            "example: move E8\n3. wall xyh - place wall in xy, h - horisontal," +
+            "v - vertical, for example: wall V7h\n4. help - print this " +
+            "helpbox\n5. quit - quit the game";
+        private const string IncorrectMessage = "Incorrect command! " +
+            "Try something else";
+        private const string ChooseColorMessage =
+            "Please, chose which player you want to play (black or white)";
+        private const string CongratulationsMessage = " Has won!";
+        private const string CurrentPlayerMessage = "Current player is ";
+        private const string DelimiterMessage = "-----------------" +
+            "----------------------------";
+        private const string SingleplayerMessage = "New Singleplayer" +
+            " Game has started!";
+        private const string MultiplayerMessage = "New Multiplayer" +
+            " Game has started!";
+        private const string FirstPlayerName = "White Player";
+        private const string SecondPlayerName = "Black Player";
+        private string _currentPlayerName = FirstPlayerName;
+
+
+        public static void WriteStartingMessage()
+        {
+            Console.WriteLine(GreetingMessage);
+            Console.WriteLine(HelpMessage);
+        }
+
+        public static void WriteIncorrectMessage() =>
+            Console.WriteLine(IncorrectMessage);
+
+        public void WriteNullOrEmptyMessage() =>
+            Console.WriteLine(NullOrEmptyMessage);
+
+        public static void WriteChooseColorMessage() =>
+            Console.WriteLine(ChooseColorMessage);
+
+        public static void WriteSingleplayerMessage() =>
+            Console.WriteLine(SingleplayerMessage);
+
+        public static void WriteMultiplayerMessage() =>
+            Console.WriteLine(MultiplayerMessage);
+
+        public void WriteHelpMessage() =>
+            Console.WriteLine(HelpMessage);
+
+        public void WriteDelimiter() =>
+            Console.WriteLine(DelimiterMessage);
+
+        public void WriteNewGame() =>
+            Console.WriteLine("To start new game, use command start with parameter");
+
+        public void WritePlayerMessage()
+        {
+            if (_currentGame.CurrentPlayer == _currentGame.FirstPlayer)
+            {
+                _currentPlayerName = FirstPlayerName;
+            }
+            else
+            {
+                _currentPlayerName = SecondPlayerName;
+            }
+
+            Console.WriteLine(CurrentPlayerMessage +
+                _currentPlayerName);
+        }
+
+        public void WriteCongratulations(IPlayer player)
+        {
+            var winner = "";
+            if (player == _currentGame.FirstPlayer)
+            {
+                winner = FirstPlayerName;
+            }
+            else
+            {
+                winner = SecondPlayerName;
+            }
+
+            Console.WriteLine(winner + CongratulationsMessage);
+        }
+
+        public void WriteBotMove(string command, Cell element)
+        {
+            string formattedCoordinates =
+                       $"{(char)(element.Coordinates.Y + 65)}" +
+                       $"{element.Coordinates.X + 1}";
+            Console.WriteLine(command + " " + formattedCoordinates);
+        }
+
+        public void WriteBotPlaceWall(string command, Wall element)
+        {
+            string formattedCoordinates =
+                        $"{(char)(element.Coordinates.Y + 83)}" +
+                        $"{element.Coordinates.X + 1}" +
+                        $"{(element.Coordinates.X == element.EndCoordinates.X ? 'v' : 'h')}";
+            Console.WriteLine(command + " " + formattedCoordinates); 
+        }
+
     }
 }
